Move compare_class_to_scores to utils.py
